♻️ refactor: 레스토랑 리스트 생성 함수 분리
