+++ src/App.js
@@ -1,8 +1,8 @@
 import Header from "./Header.js";
 import Restaurant from "./Restaurant.js";
-import { RestaurantData } from "./RestaurantData.js";
 import { categoryValue } from "./optionValue.js";
 import Component from "./Component.js";
+import createRestaurant from "./createRestaurant.js";
 class App extends Component {
   constructor($target) {
     super($target);
@@ -12,14 +12,7 @@ class App extends Component {
   render() {
     document.querySelector(".restaurant-list").innerHTML = "";
     new Header(document.querySelector(".gnb"));
-
-    RestaurantData.forEach((data) => {
-      const restaurantItem = document.createElement("li");
-      restaurantItem.classList.add("restaurant");
-      document.querySelector(".restaurant-list").appendChild(restaurantItem);
-
-      new Restaurant(restaurantItem, data);
-    });
+    createRestaurant();
   }
 }
 export default App;
